Make tailscale configuration manual :(
After attempting to automatically configure tailscale (with an auth key) in order to set up tailscaled-autoconnect (allowing for declarative configuration of extraUpFlags), it seems that since tailscale is already logged in (and does not need login or machine auth), tailscaled-autoconnect's script never runs, meaning extraUpFlags are never set. Also, this declarative setup requires a new auth key every 90 days, which is annoying.